bpf: Add ratelimiting feature and use it for ICMPv6 responses
This commit adds token bucket ratelimiting to the datapath. It is
implemented purely in BPF. A new map is added to keep track of the
buckets. A bucket can be keyed on anything, though since ICMPv6 is
currently the only user it is keyed on ifindex. The value is the current
amounts of tokens in the bucket and the last time we added tokens to the
bucket.

For every event we check if there is at least 1 token left in the bucket
if so, we decrement the token count and continue, if not we execute the
rate limiting action. Typically a timer would add new tokens into the
bucket, in our case we keep track of the last time we added tokens and
calculate how many tokens we should have added since then before we
do the token check.

This implements a burstable rate limiting mechanism. The burst size
and token refil is configurable. For ICMPv6 it is currently set to
100 replies per second with a burst size of 1000.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>