ci/cilium-config: Support passing image suffix as Action input
Make it possible to pass a custom suffix string for the Cilium
agent/operator and Hubble relay images to use in CI workflows,
defaulting to "-ci" to preserve the current behaviour.

The intended use case is to pass an empty suffix when we want to use a
release image not built as part of the CI workflows.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>