CODEOWNERS: add /pkg/iterator to sig-agent codeowner.
pkg/iterator is implements a generic iterator, as well as algorithms on
that type.

The initial use for this is to implement algorithms for processing label
vectors for metrics purposes.

However, this code is general and extendable and may likely have applications in the future in other areas.
Having it be it's own library rather than metrics specific code will
hopefully encourage code reuse and better code quality.

Thus this commits assigns it to sig-agent as a general code library
within Cilium.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>